split structure.py into structure/__init__.py and structure/symmetry.py
- rename analyze_symmetry() to get_sym_info_from_structs()
- also split corresponding unit tests
