+++ tests/structure/test_symmetry.py
@@ -5,26 +5,7 @@ from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from pymatviz.enums import Key
 
 from matbench_discovery.enums import MbdKey
-from matbench_discovery.structure import (
-    analyze_symmetry,
-    perturb_structure,
-    pred_vs_ref_struct_symmetry,
-)
-
-
-@pytest.mark.parametrize("gamma", [0.5, 1.0, 1.5, 2.0])
-def test_perturb_structure(dummy_struct: Structure, gamma: float) -> None:
-    perturbed = perturb_structure(dummy_struct, gamma=gamma)
-    assert len(perturbed) == len(dummy_struct)
-
-    for site, new in zip(dummy_struct, perturbed, strict=True):
-        assert site.specie == new.specie
-        assert tuple(site.coords) != tuple(new.coords)
-
-    # Check that different calls produce different results
-    assert perturb_structure(dummy_struct, gamma=gamma) != perturb_structure(
-        dummy_struct, gamma=gamma
-    )
+from matbench_discovery.structure import perturb_structure, symmetry
 
 
 @pytest.mark.parametrize(
@@ -42,7 +23,7 @@ def test_analyze_symmetry(
     expected_n_sym_ops: int,
 ) -> None:
     structure = request.getfixturevalue(test_structure)
-    df_sym = analyze_symmetry({"test_structure": structure})
+    df_sym = symmetry.get_sym_info_from_structs({"test_structure": structure})
 
     assert len(df_sym) == 1
     assert df_sym[Key.spg_num].iloc[0] == expected_spg_num
@@ -59,7 +40,7 @@ def test_analyze_symmetry_multiple_structures(
         "tetragonal": tetragonal_struct,
         "monoclinic": monoclinic_struct,
     }
-    df_sym = analyze_symmetry(structures)
+    df_sym = symmetry.get_sym_info_from_structs(structures)
 
     assert len(df_sym) == 3
     assert list(df_sym[Key.spg_num]) == [229, 47, 3]
@@ -70,10 +51,10 @@ def test_pred_vs_ref_struct_symmetry(
     cubic_struct: Structure, tetragonal_struct: Structure
 ) -> None:
     key = "structure"
-    df_ml = analyze_symmetry({key: cubic_struct})
-    df_dft = analyze_symmetry({key: tetragonal_struct})
+    df_ml = symmetry.get_sym_info_from_structs({key: cubic_struct})
+    df_dft = symmetry.get_sym_info_from_structs({key: tetragonal_struct})
 
-    df_compared = pred_vs_ref_struct_symmetry(
+    df_compared = symmetry.pred_vs_ref_struct_symmetry(
         df_ml, df_dft, {key: cubic_struct}, {key: tetragonal_struct}
     )
 
@@ -87,8 +68,10 @@ def test_pred_vs_ref_struct_symmetry(
 def test_analyze_symmetry_perturbed_structure(cubic_struct: Structure) -> None:
     perturbed_structure = perturb_structure(cubic_struct, gamma=1.5)
 
-    df_sym_original = analyze_symmetry({"original": cubic_struct})
-    df_sym_perturbed = analyze_symmetry({"perturbed": perturbed_structure})
+    df_sym_original = symmetry.get_sym_info_from_structs({"original": cubic_struct})
+    df_sym_perturbed = symmetry.get_sym_info_from_structs(
+        {"perturbed": perturbed_structure}
+    )
 
     assert not df_sym_original.equals(df_sym_perturbed)
     assert df_sym_perturbed[Key.spg_num].iloc[0] < df_sym_original[Key.spg_num].iloc[0]
@@ -100,8 +83,8 @@ def test_analyze_symmetry_perturbed_structure(cubic_struct: Structure) -> None:
 def test_analyze_symmetry_supercell(cubic_struct: Structure) -> None:
     supercell = cubic_struct * (2, 2, 2)
 
-    df_sym_original = analyze_symmetry({"original": cubic_struct})
-    df_sym_supercell = analyze_symmetry({"supercell": supercell})
+    df_sym_original = symmetry.get_sym_info_from_structs({"original": cubic_struct})
+    df_sym_supercell = symmetry.get_sym_info_from_structs({"supercell": supercell})
 
     assert list(df_sym_original) == list(df_sym_supercell)
     assert df_sym_original[Key.spg_num].iloc[0] == df_sym_supercell[Key.spg_num].iloc[0]
@@ -114,8 +97,8 @@ def test_pred_vs_ref_struct_symmetry_with_structures(
     df_dft = pd.DataFrame({Key.structure: [tetragonal_struct]})
 
     key = "struct1"
-    df_ml_sym = analyze_symmetry({key: cubic_struct})
-    df_dft_sym = analyze_symmetry({key: tetragonal_struct})
+    df_ml_sym = symmetry.get_sym_info_from_structs({key: cubic_struct})
+    df_dft_sym = symmetry.get_sym_info_from_structs({key: tetragonal_struct})
 
     df_ml = pd.concat([df_ml, df_ml_sym], axis=1)
     df_dft = pd.concat([df_dft, df_dft_sym], axis=1)
@@ -123,7 +106,7 @@ def test_pred_vs_ref_struct_symmetry_with_structures(
     df_ml_sym[Key.spg_num] - df_dft_sym[Key.spg_num]
 
     # must use same keys for both structures to match them in RMSD calculation
-    df_compared = pred_vs_ref_struct_symmetry(
+    df_compared = symmetry.pred_vs_ref_struct_symmetry(
         df_ml, df_dft, {key: cubic_struct}, {key: tetragonal_struct}
     )
 
@@ -152,8 +135,12 @@ def test_analyze_symmetry_primitive_vs_conventional(cubic_struct: Structure) -> 
 
     conventional_key, primitive_key = "conventional", "primitive"
     int_spg_key = "international_spg_name"
-    df_conventional = analyze_symmetry({conventional_key: cubic_struct})
-    df_primitive = analyze_symmetry({primitive_key: primitive_structure})
+    df_conventional = symmetry.get_sym_info_from_structs(
+        {conventional_key: cubic_struct}
+    )
+    df_primitive = symmetry.get_sym_info_from_structs(
+        {primitive_key: primitive_structure}
+    )
     assert df_conventional.index.name == Key.mat_id
     assert df_primitive.index.name == Key.mat_id
     assert df_primitive.index[0] == primitive_key
@@ -189,7 +176,7 @@ def test_analyze_symmetry_with_ase_atoms(cubic_struct: Structure) -> None:
     assert isinstance(atoms, Atoms)
 
     # Test with both Structure and Atoms
-    df_ase = analyze_symmetry({"ase": atoms})
+    df_ase = symmetry.get_sym_info_from_structs({"ase": atoms})
 
     assert len(df_ase) == 1
     assert df_ase.index.name == Key.mat_id
@@ -197,7 +184,7 @@ def test_analyze_symmetry_with_ase_atoms(cubic_struct: Structure) -> None:
     assert isinstance(df_ase[Key.n_sym_ops].iloc[0], int)
 
     # Test mixed dictionary of Structure and Atoms
-    df_mixed = analyze_symmetry({"pmg": cubic_struct, "ase": atoms})
+    df_mixed = symmetry.get_sym_info_from_structs({"pmg": cubic_struct, "ase": atoms})
     assert len(df_mixed) == 2
     assert df_mixed.index.tolist() == ["pmg", "ase"]
     assert all(isinstance(x, int) for x in df_mixed[Key.spg_num])
@@ -206,8 +193,8 @@ def test_analyze_symmetry_with_ase_atoms(cubic_struct: Structure) -> None:
     # should give same results for Structure and equivalent Atoms
     cubic_atoms = cubic_struct.to_ase_atoms()
 
-    df_struct = analyze_symmetry({"struct": cubic_struct})
-    df_atoms = analyze_symmetry({"atoms": cubic_atoms})
+    df_struct = symmetry.get_sym_info_from_structs({"struct": cubic_struct})
+    df_atoms = symmetry.get_sym_info_from_structs({"atoms": cubic_atoms})
 
     # Reset index to avoid index name mismatch in frame comparison
     df_struct.index = df_atoms.index
